reorganized folder structure & worked on some Login & Register page
